backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
supabase
python-dotenv
apscheduler
python-jose[cryptography]
passlib[bcrypt]